agents/honcho_fact_memory/pyproject.toml
[tool.poetry]
name = "honcho-fact-memory"
version = "0.1.0"
description = ""
authors = ["vintro <vince@plasticlabs.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
realtime = "^1.0.2"
supabase = "^2.4.0"
langchain-openai = "^0.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
